Expose setNeedsSafetyTraining
